requirements/requirements.txt
Django
South>=1.0
argparse
django-admin-honeypot
django-axes
django-dbbackup
django-bootstrap-pagination
django-bootstrap-toolkit
django-bootstrap3
django-braces
django-cache-utils
django-chroniker
django-cities-light
django-clear-cache
django-countries
django-country-blocker
#django-country-dialcode
django-currencies
django-database-size
django-datetime-widget
django-google-charts
django-iban
django-import-export
django-jsonfield
django-qsstats-magic
django-queryset-csv
django-mass-edit
django-mathfilters
django-model-report
django-request
django-secure
django-simple-import
django-system-monitor
django-tables2
django-templated-email
GeoIP
netaddr
netifaces
psutil
psycopg2
python-memcached
raven
requests
pytz
suds
vatnumber
wsgiref
